requirements.txt
transformers
torch
chromadb
python-telegram-bot
sentence-transformers